feat: move implementation to repository
