+++ app/Http/Controllers/Api/SchoolController.php
@@ -20,29 +20,29 @@ class SchoolController extends ApiController
     }
 
     public function show(Request $request, $id) {
-        $school = $this->service()->school($request->user(), $id);
+        $school = $this->service()->repo()->school($request->user(), $id);
         $this->authorize('view', $school); /** ensure the current user has view rights */
         return $school;
     }
 
     public function index(Request $request, SchoolFilters $filters) {
-        return $this->service()->schools($request->user(), $filters);
+        return $this->service()->repo()->schools($request->user(), $filters);
     }
 
     public function destroy(Request $request, $id) {
-        $school = $this->service()->school($request->user(), $id);
+        $school = $this->service()->repo()->school($request->user(), $id);
         $this->authorize('delete', $school); /** ensure the current user has delete rights */
-        $this->service()->delete($request->user(), $id);
+        $this->service()->repo()->delete($request->user(), $id);
         return $this->ok();
     }
 
     public function store(SchoolRequest $request) {
-        $school = $this->service()->create($request->user(), $request->all());
+        $school = $this->service()->repo()->create($request->user(), $request->all());
         return $this->json($school);
     }
 
     public function update(Request $request, $id) {
-        $school = $this->service()->update($request->user(), $id, $request->all());
+        $school = $this->service()->repo()->update($request->user(), $id, $request->all());
         $this->authorize('update', $school);
         return $this->json($school);
     }
